requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LLM Providers
boto3>=1.34.0  # AWS Bedrock
anthropic>=0.8.0  # Claude (future)
openai>=1.6.0  # OpenAI (future)

# Vector Database
chromadb>=0.4.22

# Redis (Session Management)
redis>=5.0.0
aioredis>=2.0.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.11.0

# Text Processing
langchain>=0.1.0
langchain-community>=0.0.10
tiktoken>=0.5.2

# Terraform HCL Parsing
python-hcl2>=4.3.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
aiofiles>=23.2.1
httpx>=0.25.0
orjson>=3.9.0
